+++ example/Nix/nixos-host/configuration.nix
@@ -4,7 +4,7 @@
 ##                                                                            ##
 ################################################################################
 
-{ pkgs, ... }:
+{ inputs, pkgs, ... }:
 
 {
   # This value determines the NixOS release with which your system is to be
@@ -16,14 +16,15 @@
   imports = [
     # Import the confkit NixOS module to get ready-to-use configurations for
     # several tools.
-    ../../confkit/nixos
+    inputs.confkit.nixosModules.confkit-nixos
 
     # Import the home-manager NixOS module to handle user configurations
     # declaratively.
-    ../../home-manager/nixos
+    inputs.home-manager.nixosModules.home-manager
 
     # Include the results of the hardware scan.
-    ./hardware-configuration.nix
+    # TODO: Uncomment this once it has been generated.
+    # ./hardware-configuration.nix
 
     # Configuration for the users.
     ./users/root
@@ -37,6 +38,7 @@
   confkit = {
     info = {
       name = "nixos-host";
+      # TODO: Change the machineId.
       machineId = "c6dc57dbf4e9384215c6d0e6616d2ff2";
       location = "kerguelen";
     };
@@ -102,6 +104,7 @@
 
   services = {
     avahi = { enable = true; nssmdns = true; };
+    tlp.enable = false;
 
     redshift = {
       enable = true;
@@ -147,17 +150,17 @@
     openssl
     pandoc
     redshift-plasma-applet
-    wine
 
     # TeXLive can be useful for tools like Pandoc or Org.
     texlive.combined.scheme-full
 
     # Applications
+    element-desktop
     firefox
     gimp
     gwenview
     kate
-    keepassx2
+    keepassxc
     konversation
     libreoffice
     mpv